Updated ATS-EcoSIM fork with ATS-1.6
The versions should now be compatible. Some minor edits needed to be
made outside of the merge, due to capitalization changes of some
header files, and renaming of the requireAtCurrent/Next functions.

This code compiles, but currently is not running due to some
changes to transport in the input files.